add categorys handle the stock when a bill is submitted (javanullpointerexception to resolve but the stock is correctly updated in the debug)

+++ ProjAppliWeb/src/main/java/m2/project/model/TVA.java
@@ -40,7 +40,7 @@ public class TVA {
 		return tva;
 	}
 
-	public void setTva(int tva) {
+	public void setTva(double tva) {
 		this.tva = tva;
 	}
